retarget to iCE40 UltraPlus Breakout Board
Note that you will need a separate ftdi board for OPTO mode or
will need to bodge the connections from the onboard ftdi's port
B.

Also, the board has 2 fewer LEDs than the iCEPick board, so you
need to add those yourself.

The iCE40UP also only has 1 PLL, and it can't be used if you use
pin 35 to bring in the clock. Instead, remove jumper J51 and
connect 45A_G1 to ICE_CLK on header B. (can be done with a
jumper)